requirements.txt
pydub
python-magic